gpu: started refactor
input: Changed digitalbutton threshold to lower value
quiet warnings

+++ src/renderer/opengl/shader/Attribute.cpp
@@ -8,7 +8,7 @@ void Attribute::enable() { glEnableVertexAttribArray(id); }
 
 void Attribute::disable() { glDisableVertexAttribArray(id); }
 
-void Attribute::pointer(GLint size, GLenum type, GLsizei stride, int pointer) {
+void Attribute::pointer(GLint size, GLenum type, GLsizei stride, uintptr_t pointer) {
     enable();
     if (type == GL_INT || type == GL_UNSIGNED_INT)
         glVertexAttribIPointer(id, size, type, stride, (const GLvoid*)pointer);
